Add files via upload
in this project we can add delete update to a database dynamically from this GUI the insert will be possible by pressing the save button, clicking on a row, and changing data will update the data on save. delete will delete the row from the database and from the table. you can also edit the data directly from the Jtable on the right, there are constraints that are applied on both panels and will give a dialogue box, ex name cannot have numbers in them